% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_kernel.R
\name{exponential_kernel}
\alias{exponential_kernel}
\title{Exponential Kernel}
\usage{
exponential_kernel(l = 1)
}
\arguments{
\item{l}{Length scale.}
}
\value{
matrix
}
\description{
The exponential kernel on a pair of matrices.
The exponential kernel is defined as:
\deqn{ k(x1, x2) = e^( -(||x1 - x2|| / l)) }
}
\examples{
k <- exponential_kernel(l = 3)
k(rnorm(4), rnorm(4, 1, 2))
}
