% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_kernel.R
\name{linear_kernel}
\alias{linear_kernel}
\title{Linear Kernel}
\usage{
linear_kernel(sigma = 0)
}
\arguments{
\item{sigma}{homogeneity parameter, scalar.}
}
\value{
matrix
}
\description{
The linear kernel on a pair of matrices.
The linear kernel is defined as:
\deqn{ k(x1, x2) = \sigma^2 + dot(x1, x2)}
}
\examples{
k <- linear_kernel(sigma = 3)
k(rnorm(4), rnorm(4, 1, 2))
}
