% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern_kernel.R
\name{matern_kernel}
\alias{matern_kernel}
\title{Matérn Kernel}
\usage{
matern_kernel(l = 1, nu = 2.5)
}
\arguments{
\item{l}{Length scale.}

\item{nu}{smoothness parameter, positive scalar.}
}
\value{
matrix
}
\description{
The Matérn kernel on a pair of matrices.
}
\examples{
k <- exponential_kernel(l = 3)
k(rnorm(4), rnorm(4, 1, 2))
}
