% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_kernel.R
\name{new_kernel}
\alias{new_kernel}
\title{Construct a New Kernel Constructor}
\usage{
new_kernel(expr, name, ...)
}
\arguments{
\item{expr}{}

\item{name}{}

\item{...}{}
}
\value{
function
}
\description{
Construct a New Kernel Constructor
}
