% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset_parameters.R
\name{parameters}
\alias{parameters}
\alias{get_par}
\alias{get_par.seoro_kernel}
\title{Get and Set Kernel Parameters}
\usage{
get_par(k)

\method{get_par}{seoro_kernel}(k)
}
\arguments{
\item{k}{a kernel}
}
\description{
Get and Set Kernel Parameters
}
