% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_kernel.R
\name{periodic_kernel}
\alias{periodic_kernel}
\title{Basic Periodic Kernel}
\usage{
periodic_kernel(l = 1, omega = 1)
}
\arguments{
\item{l}{Length scale.}

\item{omega}{period parameter, scalar.}
}
\value{
matrix
}
\description{
The basic periodic kernel on a pair of matrices.
}
\examples{
k <- periodic_kernel(l = 3, omega = 2 * pi)
k(rnorm(4), rnorm(4, 1, 2))
}
