% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_math.R
\name{+.seoro_kernel}
\alias{+.seoro_kernel}
\title{Add Two Kernels}
\usage{
\method{+}{seoro_kernel}(l, r)
}
\arguments{
\item{l}{kernel}

\item{r}{kernel}
}
\value{
kernel
}
\description{
Add Two Kernels
}
