% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_kernel.R
\name{polynomial_kernel}
\alias{polynomial_kernel}
\title{Polynomial Kernel}
\usage{
polynomial_kernel(sigma = 0, nu = 2)
}
\arguments{
\item{sigma}{homogeneity parameter, scalar.}

\item{nu}{degree parameter, positive integer}
}
\value{
matrix
}
\description{
The polynomial kernel on a pair of matrices.
The plonomial kernel is defined as:
\deqn{ k(x1, x2) = (\sigma^2 + dot(x1, x2))^\nu}
}
\examples{
k <- polynomial_kernel(sigma = 3)
k(rnorm(4), rnorm(4, 1, 2))
}
