% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf_kernel.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{Radial Basis Function Kernel}
\usage{
rbf_kernel(x1, x2, l = 1)
}
\arguments{
\item{x1}{First numeric vector of length d}

\item{x2}{Second numeric vector of length d}

\item{l}{Length scale. Scalar or numeric vector of length d.}
}
\value{
scalar
}
\description{
Radial Basis Function Kernel
}
\examples{
rbf_kernel(rnorm(4), rnorm(4, 1, 2), l = 3)
}
