% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqdist.R
\name{sqdist}
\alias{sqdist}
\title{Matrix of Squared Euclidean Distances}
\usage{
sqdist(x1, x2, tolerance = 9)
}
\arguments{
\item{x1}{matrix}

\item{x2}{matrix}

\item{tolerance}{integer, number of digits. The distances are rounded to this
tolerance primarily to ensure that the distance between two identical vectors
is exactly zero.}
}
\value{
matrix
}
\description{
The matrix of squared 2-norm distances between two matrices.
}
\examples{
sqdist(matrix(rnorm(4), 2, 2), matrix(rnorm(10), 5, 2))
}
