% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_noise_kernel.R
\name{white_noise_kernel}
\alias{white_noise_kernel}
\title{White Noise Kernel}
\usage{
white_noise_kernel(sigma = 1)
}
\arguments{
\item{sigma}{scale parameter, scalar.}
}
\value{
matrix
}
\description{
The white noise aka the white noise kernel on a pair of matrices.
The white noise kernel is defined as:
\deqn{ \sigma^2\delta_{ij} }
}
\examples{
k <- white_noise_kernel(sigma = 2)
k(rnorm(4), rnorm(4, 1, 2))
}
