% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_kernel.R
\name{new_kernel}
\alias{new_kernel}
\title{Construct a New Kernel Constructor}
\usage{
new_kernel(expr, name, ...)
}
\arguments{
\item{expr}{kernel expression that calculates the covariance matrix given two
input matrices x1 and x2.}

\item{name}{name of the kernel}

\item{...}{default values for any parameters in the expression. Arguments
must be named.}
}
\value{
function
}
\description{
Construct a New Kernel Constructor
}
