% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf_kernel.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{Radial Basis Function Kernel}
\usage{
rbf_kernel(l = 1)
}
\arguments{
\item{l}{Length scale. Scalar or numeric vector of length d.}
}
\value{
scalar
}
\description{
The Radial Basis Function (RBF) kernel on a pair of vectors.
The RBF kernel is defined as:
\deqn{ k(x1, x2) = e^( -0.5(||x1 - x2|| / l)^2) }
}
\examples{
k <- rbf_kernel(l = 3)
k(rnorm(4), rnorm(4, 1, 2))
}
